backend/requirements.txt
flask
flask-cors
scikit-learn
numpy
python-dotenv
